Create ReminderFragment and recreate same funcionality in ReminderFragment from ReminderActivity
