+++ app/src/main/java/com/example/avjindersinghsekhon/minimaltodo/Settings/SettingsFragment.java
@@ -7,9 +7,9 @@ import android.preference.CheckBoxPreference;
 import android.preference.PreferenceFragment;
 
 import com.example.avjindersinghsekhon.minimaltodo.Analytics.AnalyticsApplication;
-import com.example.avjindersinghsekhon.minimaltodo.Main.MainActivity;
-import com.example.avjindersinghsekhon.minimaltodo.Utility.PreferenceKeys;
+import com.example.avjindersinghsekhon.minimaltodo.Main.MainFragment;
 import com.example.avjindersinghsekhon.minimaltodo.R;
+import com.example.avjindersinghsekhon.minimaltodo.Utility.PreferenceKeys;
 
 public class SettingsFragment extends PreferenceFragment implements SharedPreferences.OnSharedPreferenceChangeListener{
     AnalyticsApplication app;
@@ -26,19 +26,19 @@ public class SettingsFragment extends PreferenceFragment implements SharedPrefer
     public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
         PreferenceKeys preferenceKeys = new PreferenceKeys(getResources());
         if(key.equals(preferenceKeys.night_mode_pref_key)){
-            SharedPreferences themePreferences = getActivity().getSharedPreferences(MainActivity.THEME_PREFERENCES, Context.MODE_PRIVATE);
+            SharedPreferences themePreferences = getActivity().getSharedPreferences(MainFragment.THEME_PREFERENCES, Context.MODE_PRIVATE);
             SharedPreferences.Editor themeEditor = themePreferences.edit();
             //We tell our MainLayout to recreate itself because mode has changed
-            themeEditor.putBoolean(MainActivity.RECREATE_ACTIVITY, true);
+            themeEditor.putBoolean(MainFragment.RECREATE_ACTIVITY, true);
 
             CheckBoxPreference checkBoxPreference = (CheckBoxPreference)findPreference(preferenceKeys.night_mode_pref_key);
             if(checkBoxPreference.isChecked()){
                 //Comment out this line if not using Google Analytics
                 app.send(this, "Settings", "Night Mode used");
-                themeEditor.putString(MainActivity.THEME_SAVED, MainActivity.DARKTHEME);
+                themeEditor.putString(MainFragment.THEME_SAVED, MainFragment.DARKTHEME);
             }
             else{
-                themeEditor.putString(MainActivity.THEME_SAVED, MainActivity.LIGHTTHEME);
+                themeEditor.putString(MainFragment.THEME_SAVED, MainFragment.LIGHTTHEME);
             }
             themeEditor.apply();
